+++ scripts/run_ptyrad.py
@@ -8,7 +8,7 @@ PATH_TO_PTYRAD = "/home/fs01/cl2696/workspace/ptyrad"  # Change this for the pty
 sys.path.append(PATH_TO_PTYRAD)
 from ptyrad.data_io import load_params  # noqa: E402
 from ptyrad.reconstruction import PtyRADSolver  # noqa: E402
-from ptyrad.utils import print_system_info, set_gpu_device  # noqa: E402
+from ptyrad.utils import print_system_info, set_accelerator, set_gpu_device # noqa: E402
 
 if __name__ == "__main__":
     # If you want to run with GPU 0
@@ -24,11 +24,14 @@ if __name__ == "__main__":
     parser.add_argument("--gpuid",       type=int, required=False, default=None)
     args = parser.parse_args()
     
+    # Set up accelerator for multiGPU/mixed-precision setting, note that thess has no effect when we launch it with just `python <script>`
+    accelerator = set_accelerator()
+        
     print_system_info()
     params = load_params(args.params_path)
     device = set_gpu_device(args.gpuid)
     
-    ptycho_solver = PtyRADSolver(params, device=device)
+    ptycho_solver = PtyRADSolver(params, device=device, acc=accelerator)
 
     ptycho_solver.run()